extract deploy logic from config.js
